Gjør Testcontainers oppsett til ikke abstrakt
