+++ src/test/kotlin/no/nav/helse/flex/testconfig/IntegrasjonTestOppsett.kt
@@ -17,6 +17,8 @@ import org.springframework.boot.test.autoconfigure.web.servlet.MockMvcPrint
 import org.springframework.boot.test.context.SpringBootTest
 import org.springframework.kafka.config.KafkaListenerEndpointRegistry
 import org.springframework.kafka.test.utils.ContainerTestUtils
+import org.springframework.test.context.DynamicPropertyRegistry
+import org.springframework.test.context.DynamicPropertySource
 import org.springframework.test.web.servlet.MockMvc
 
 @TestInstance(TestInstance.Lifecycle.PER_CLASS)
@@ -31,7 +33,7 @@ import org.springframework.test.web.servlet.MockMvc
     ],
 )
 @AutoConfigureMockMvc(print = MockMvcPrint.NONE, printOnlyOnFailure = false)
-abstract class IntegrasjonTestOppsett : TestcontainersTestOppsett() {
+abstract class IntegrasjonTestOppsett {
     @Autowired
     lateinit var mockMvc: MockMvc
 
@@ -52,6 +54,14 @@ abstract class IntegrasjonTestOppsett : TestcontainersTestOppsett() {
 
     @Autowired
     lateinit var kafkaListenerEndpointRegistry: KafkaListenerEndpointRegistry
+
+    companion object {
+        @JvmStatic
+        @DynamicPropertySource
+        fun dynamicProperties(registry: DynamicPropertyRegistry) {
+            TestcontainersOppsett.setPropertiesInContext(registry)
+        }
+    }
 
     @BeforeAll
     fun beforeAllFelles() {
